+++ kernels/moe/moe_gemm_2stage/gemm2.py
@@ -33,8 +33,8 @@ from flydsl._mlir.dialects import llvm, scf
 from flydsl.expr.typing import T
 from kernels.common.kernels_common import _if_then
 from kernels.common.mem_ops import buffer_atomic_add
-from kernels.mma.mfma_epilogues import c_shuffle_epilog, default_epilog
-from kernels.mma.mfma_preshuffle_pipeline import (
+from kernels.common.mma.mfma_epilogues import c_shuffle_epilog, default_epilog
+from kernels.common.mma.mfma_preshuffle_pipeline import (
     buffer_copy_gmem16_dwordx4,
     extract_bf16_scale,
     lds_store_4b_xor16,
